requirements.txt
# Core dependencies
click>=8.1.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Kubernetes
kubernetes>=28.0.0

# FastAPI (for web service)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Template engine
jinja2>=3.1.0

# Infrastructure
pulumi>=3.0.0,<4.0.0
pulumi-gcp>=7.0.0,<8.0.0
pulumi-kubernetes>=4.0.0,<5.0.0

